Parking Lot
Added the Parking Lot. Added the first zombie encounter. Added a combat
system for the axe. Fixed a bug where Inform would keep on reporting
the zombie was lying on the floor when he was actually attacking you.
Fixed a problem where just attacking the person would say "Violence
isn't the answer to this."