Engine Room
Added the Engine Room on the Engine Car on the Train. Added the panel
and the missing lever. Nasty bugs such as the engine room dissappearing
and the levers not being properly named made this process of coding
very slow. Next commit I will have coded all the lever must be turned
on and the button pushed for the train to start, and the coupler
detached, for the train to start.